Fixed flaky test in 'apache/pinot' repository. (#738)
Contributing fix for 'CrcUtilsTest' flaky test(#10442), PR is created. This issue is confirmed by one of the PMCs of this project. Link to the PR: apache/pinot#10445.

The following changes were made to pr-data.csv in this PR:

Flaky test org.apache.pinot.core.util.CrcUtilsTest.test1 was already in pr-data.csv, I updated the line for this flaky test to include the fix PR link and set the status to opened.