Add prStatusCheck.py for automated PR status updates
This script automates checking and updating PR statuses in the iDFlakies dataset( {gr/pr/py}-data.csv ). It queries PR status via the GitHub API and will update tests' status from "Opened" to "Accepted" when PRs are merged, and flagging closed-but-not-merged PRs as "Unknown" for manual review.

It also supports per-file row-range filtering and can ignore tests based on their names.

Currently I can update the status of all pull requests in the dataset within approximately three minutes using this script and maintain a very high level of accuracy (which appears to be 100% so far, at least I haven’t seen a misclassification yet).

For a more detailed and readable explanation, see:
https://github.com/KangweiZhu/idoft/blob/better-updater/auto-update-dataset/python/README.md#prstatuscheckpy